Refactor AddListenModal component
move to functional component and usage with niceModal library
